requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
werkzeug==2.3.6
pymongo==4.4.1
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
fpdf==1.7.2
